DbRepository - AttributeAccessors trait
